refactor(tests,infra): split integration.rs by domain; migrate subcmd; x402 solana feature gate
