+++ crates/gateway/tests/integration_payments_402.rs
@@ -1,0 +1,389 @@
+//! Integration tests — payments_402.
+//!
+//! Split from the original `tests/integration.rs`.
+//! Shared helpers live in `tests/common/mod.rs`.
+
+#[path = "common/mod.rs"]
+mod common;
+
+#[allow(unused_imports)]
+use std::collections::HashMap;
+#[allow(unused_imports)]
+use std::pin::Pin;
+use std::sync::Arc;
+
+#[allow(unused_imports)]
+use async_trait::async_trait;
+use axum::body::Body;
+use axum::http::{Request, StatusCode};
+use base64::Engine;
+#[allow(unused_imports)]
+use futures::stream;
+use http_body_util::BodyExt;
+#[allow(unused_imports)]
+use tokio::sync::RwLock;
+use tower::ServiceExt;
+
+#[allow(unused_imports)]
+use gateway::config::AppConfig;
+use gateway::middleware::rate_limit::{RateLimitConfig, RateLimiter};
+#[allow(unused_imports)]
+use gateway::providers::health::{CircuitBreakerConfig, ProviderHealthTracker};
+#[allow(unused_imports)]
+use gateway::providers::{ChatStream, LLMProvider, ProviderRegistry};
+#[allow(unused_imports)]
+use gateway::services::ServiceRegistry;
+#[allow(unused_imports)]
+use gateway::{build_router, AppState};
+#[allow(unused_imports)]
+use solvela_protocol::{
+    ChatChoice, ChatChunk, ChatChunkChoice, ChatDelta, ChatMessage, ChatResponse, ModelInfo, Role,
+    Usage,
+};
+#[allow(unused_imports)]
+use solvela_router::models::ModelRegistry;
+#[allow(unused_imports)]
+use solvela_x402::traits::{Error as X402Error, PaymentVerifier};
+#[allow(unused_imports)]
+use solvela_x402::types::{
+    EscrowPayload, PayloadData, PaymentAccept, PaymentPayload, Resource, SettlementResult,
+    SolanaPayload, VerificationResult, SOLANA_NETWORK, USDC_MINT,
+};
+
+#[allow(unused_imports)]
+use common::*;
+
+#[tokio::test]
+async fn test_chat_returns_402_without_payment() {
+    let app = test_app();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    // Should return 402 Payment Required
+    assert_eq!(response.status(), StatusCode::PAYMENT_REQUIRED);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+
+    // The error message contains the serialized PaymentRequired JSON
+    let error_msg = json["error"]["message"].as_str().unwrap();
+    let payment_info: serde_json::Value = serde_json::from_str(error_msg).unwrap();
+
+    assert_eq!(payment_info["x402_version"], 2);
+    assert!(payment_info["accepts"].is_array());
+    assert!(payment_info["cost_breakdown"]["total"].is_string());
+    assert_eq!(payment_info["cost_breakdown"]["currency"], "USDC");
+    assert_eq!(payment_info["cost_breakdown"]["fee_percent"], 5);
+}
+
+#[tokio::test]
+async fn test_chat_with_payment_returns_mock_response() {
+    let app = test_app_with_mock_provider();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header(
+                    "payment-signature",
+                    valid_payment_header("/v1/chat/completions"),
+                )
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+    assert_eq!(json["object"], "chat.completion");
+    assert_eq!(json["choices"][0]["message"]["content"], "[mock response]");
+    assert!(json["usage"]["total_tokens"].is_number());
+}
+
+/// Paid requests with NO provider configured should return 500 (stub rejection).
+#[tokio::test]
+async fn test_chat_paid_no_provider_returns_500() {
+    let app = test_app();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header(
+                    "payment-signature",
+                    valid_payment_header("/v1/chat/completions"),
+                )
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::INTERNAL_SERVER_ERROR);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+    assert_eq!(json["error"]["type"], "internal_error");
+}
+
+#[tokio::test]
+async fn test_malformed_payment_header_returns_402() {
+    let app = test_app();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header("payment-signature", "fake-payment-for-testing")
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    // Malformed (non-decodable) payment headers must be rejected — never served free
+    assert_eq!(response.status(), StatusCode::PAYMENT_REQUIRED);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+    assert_eq!(json["error"]["type"], "invalid_payment");
+    assert!(json["error"]["message"]
+        .as_str()
+        .unwrap()
+        .contains("could not be decoded"));
+}
+
+#[tokio::test]
+async fn test_402_response_contains_x402_fields() {
+    let app = test_app();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Tell me about Solana."}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::PAYMENT_REQUIRED);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+
+    // Parse the embedded PaymentRequired JSON from the error message
+    let error_msg = json["error"]["message"].as_str().unwrap();
+    let pr: serde_json::Value = serde_json::from_str(error_msg).unwrap();
+
+    // x402 version
+    assert_eq!(pr["x402_version"], 2);
+
+    // accepts array with Solana network
+    let accepts = pr["accepts"].as_array().unwrap();
+    assert!(!accepts.is_empty());
+    assert!(accepts[0]["network"]
+        .as_str()
+        .unwrap()
+        .starts_with("solana:"));
+    assert_eq!(accepts[0]["scheme"], "exact");
+    assert_eq!(
+        accepts[0]["asset"],
+        "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
+    );
+    assert!(accepts[0]["amount"].is_string());
+    assert!(accepts[0]["max_timeout_seconds"].is_number());
+
+    // cost breakdown fields
+    let cb = &pr["cost_breakdown"];
+    assert!(cb["provider_cost"].is_string());
+    assert!(cb["platform_fee"].is_string());
+    assert!(cb["total"].is_string());
+    assert_eq!(cb["currency"], "USDC");
+    assert_eq!(cb["fee_percent"], 5);
+
+    // total should be > 0
+    let total: f64 = cb["total"].as_str().unwrap().parse().unwrap();
+    assert!(total > 0.0, "total cost should be positive");
+
+    // platform_fee should be ~5% of provider_cost
+    let provider_cost: f64 = cb["provider_cost"].as_str().unwrap().parse().unwrap();
+    let platform_fee: f64 = cb["platform_fee"].as_str().unwrap().parse().unwrap();
+    let expected_fee = provider_cost * 0.05;
+    assert!(
+        (platform_fee - expected_fee).abs() < 0.000001,
+        "platform fee {platform_fee} should be ~5% of provider cost {provider_cost}"
+    );
+}
+
+#[tokio::test]
+async fn test_chat_with_base64_payment_header() {
+    let app = test_app_with_mock_provider();
+
+    // Build a valid PaymentPayload and base64-encode it
+    let payment_payload = PaymentPayload {
+        x402_version: 2,
+        resource: Resource {
+            url: "/v1/chat/completions".to_string(),
+            method: "POST".to_string(),
+        },
+        accepted: PaymentAccept {
+            scheme: "exact".to_string(),
+            network: SOLANA_NETWORK.to_string(),
+            amount: TEST_PAYMENT_AMOUNT.to_string(),
+            asset: USDC_MINT.to_string(),
+            pay_to: TEST_RECIPIENT_WALLET.to_string(),
+            max_timeout_seconds: 300,
+            escrow_program_id: None,
+        },
+        payload: PayloadData::Direct(SolanaPayload {
+            transaction: "dGVzdHRyYW5zYWN0aW9u".to_string(), // base64("testtransaction")
+        }),
+    };
+
+    let json_bytes = serde_json::to_vec(&payment_payload).unwrap();
+    let encoded = base64::engine::general_purpose::STANDARD.encode(&json_bytes);
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header("payment-signature", encoded)
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    // Base64-encoded payment should be successfully decoded and verified
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+    assert_eq!(json["object"], "chat.completion");
+    assert_eq!(json["choices"][0]["message"]["content"], "[mock response]");
+}
+
+#[tokio::test]
+async fn test_payment_status_none_on_402() {
+    let app = test_app();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header("x-rcr-debug", "true")
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    // 402 responses go through GatewayError, not the handler's debug header path.
+    // But request ID should still be present (middleware handles it).
+    assert_eq!(response.status(), StatusCode::PAYMENT_REQUIRED);
+    assert!(response.headers().get("x-rcr-request-id").is_some());
+}
+
+/// G.2 Test 12: Payment verified status on paid requests.
+///
+/// A properly-paid request passes the AlwaysPassVerifier and the provider
+/// responds successfully. Debug headers should show PaymentStatus::Verified.
+#[tokio::test]
+async fn test_payment_verified_reaches_provider_path() {
+    let app = test_app_with_mock_provider();
+
+    let body = serde_json::json!({
+        "model": "openai/gpt-4o",
+        "messages": [{"role": "user", "content": "Hello!"}],
+    });
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .method("POST")
+                .uri("/v1/chat/completions")
+                .header("content-type", "application/json")
+                .header("x-rcr-debug", "true")
+                .header(
+                    "payment-signature",
+                    valid_payment_header("/v1/chat/completions"),
+                )
+                .body(Body::from(serde_json::to_vec(&body).unwrap()))
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+    assert!(response.headers().get("x-rcr-request-id").is_some());
+    assert_eq!(
+        response
+            .headers()
+            .get("x-rcr-payment-status")
+            .expect("payment status debug header must be present")
+            .to_str()
+            .unwrap(),
+        "verified"
+    );
+}
