+++ crates/x402/Cargo.toml
@@ -11,24 +11,36 @@ keywords = ["x402", "solana", "usdc", "payments", "escrow"]
 categories = ["api-bindings", "cryptography"]
 
 [dependencies]
+# Core (chain-agnostic) deps — needed by `traits`, `types`, `facilitator`.
 solvela-protocol = { workspace = true }
-
 serde = { workspace = true }
 serde_json = { workspace = true }
 async-trait = { workspace = true }
 thiserror = { workspace = true }
 base64 = { workspace = true }
-bs58 = { workspace = true }
 tracing = { workspace = true }
-metrics = { workspace = true }
-reqwest = { workspace = true }
 tokio = { workspace = true }
-sha2 = { workspace = true }
-ed25519-dalek = { workspace = true }
-curve25519-dalek = { workspace = true }
-zeroize = { workspace = true }
+
+# Solana-specific deps — gated behind the `solana` feature so EVM-only or
+# other-chain consumers can compile this crate without the Solana toolchain.
+bs58 = { workspace = true, optional = true }
+sha2 = { workspace = true, optional = true }
+ed25519-dalek = { workspace = true, optional = true }
+curve25519-dalek = { workspace = true, optional = true }
+zeroize = { workspace = true, optional = true }
+reqwest = { workspace = true, optional = true }
+metrics = { workspace = true, optional = true }
 sqlx = { workspace = true, optional = true }
 
 [features]
-default = []
-postgres = ["sqlx"]
+default = ["solana"]
+solana = [
+    "dep:bs58",
+    "dep:sha2",
+    "dep:ed25519-dalek",
+    "dep:curve25519-dalek",
+    "dep:zeroize",
+    "dep:reqwest",
+    "dep:metrics",
+]
+postgres = ["sqlx", "solana"]
